don't try to do a query if there is nothing to change

+++ tools/X_remove_missing_files.R
@@ -28,23 +28,26 @@ file_tbl %<>% mutate(file_found = file_exists(path))
 
 bad_file_md5 <- file_tbl %>% filter(!file_found) %>% pull(file_md5)
 
+if(length(bad_file_md5)>0){
+  tables_to_prune <- c("std_stat_data", "file_schedule", "file_info", "cont_data", "files", "files_ignore")
+  
+  
+  cmd <- expand.grid(bad_file_md5, tables_to_prune, stringsAsFactors = FALSE)
+  
+  
+  sql_query <- paste0("DELETE FROM ",cmd$Var2," WHERE file_md5='",cmd$Var1,"';")
+  
+  
+  con <- poolCheckout(pool)
+        dbBegin(con)
+        
+        
+  for(i in seq_along(sql_query)){
+    dbSendQuery(con,sql_query[i])
+    dbCommit(con)
+  }
+  
+  poolReturn(con)
+  
 
-tables_to_prune <- c("std_stat_data", "file_schedule", "file_info", "cont_data", "files")
-
-
-cmd <- expand.grid(bad_file_md5, tables_to_prune, stringsAsFactors = FALSE)
-
-
-sql_query <- paste0("DELETE FROM ",cmd$Var2," WHERE file_md5='",cmd$Var1,"';")
-
-
-con <- poolCheckout(pool)
-      dbBegin(con)
-      
-      
-for(i in seq_along(sql_query)){
-  dbSendQuery(con,sql_query[i])
-  dbCommit(con)
 }
-
-poolReturn(con)
